remove enum validation for response models
