+++ incydr/enums/file_events.py
@@ -454,21 +454,3 @@ class RemoteActivity(_Enum):
 class Shared(_Enum):
     FALSE = "FALSE"
     TRUE = "TRUE"
-
-
-class SearchProblemType(_Enum):
-    ILLEGAL_GROUP = "ILLEGAL_GROUP"
-    ILLEGAL_OPERATOR = "ILLEGAL_OPERATOR"
-    ILLEGAL_SORT = "ILLEGAL_SORT"
-    ILLEGAL_VALUE = "ILLEGAL_VALUE"
-    INVALID_PAGE_NUMBER = "INVALID_PAGE_NUMBER"
-    INVALID_PAGE_SIZE = "INVALID_PAGE_SIZE"
-    INVALID_PAGE_TOKEN = "INVALID_PAGE_TOKEN"
-    MAX_FILTERS_EXCEEDED = "MAX_FILTERS_EXCEEDED"
-    MAX_LENGTH_EXCEEDED = "MAX_LENGTH_EXCEEDED"
-    MISSING_FILTERS = "MISSING_FILTERS"
-    MISSING_GROUPS = "MISSING_GROUPS"
-    MISSING_VALUE = "MISSING_VALUE"
-    QUERY_TIMEOUT = "QUERY_TIMEOUT"
-    SEARCH_FAILED = "SEARCH_FAILED"
-    UNEXPECTED_VALUE = "UNEXPECTED_VALUE"
